refactor(a): faz estrutura dentro de PanelSelector conformar com nova organização de código
