+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/index.tsx
@@ -1,0 +1,1 @@
+export { ListContainer as List } from "./container";
