+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable/modules/EmptyTableMessage.tsx
