+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/modules/print/diabetes/print.adapter.ts
